TINKERPOP-1644 Minor fixup of variable assignment

+++ gremlin-groovy/src/main/java/org/apache/tinkerpop/gremlin/groovy/jsr223/customizer/CompilationOptionsCustomizerProvider.java
@@ -38,7 +38,7 @@ public class CompilationOptionsCustomizerProvider implements CompilerCustomizerP
     }
 
     public CompilationOptionsCustomizerProvider(final Long expectedCompilationTime) {
-        this.expectedCompilationTime = expectedCompilationTime.intValue();
+        this.expectedCompilationTime = expectedCompilationTime;
     }
 
     public long getExpectedCompilationTime() {
